project-manager-api/src/main/java/com/project/manager/convert/StatusConvert.java
package com.project.manager.convert;

import com.project.manager.constant.StatusEnum;
import jakarta.persistence.AttributeConverter;
import jakarta.persistence.Converter;

import java.util.stream.Stream;

@Converter(autoApply = true)
public class StatusConvert implements AttributeConverter<StatusEnum, String> {

    @Override
    public String convertToDatabaseColumn(StatusEnum status) {
        return status.getCode();
    }

    @Override
    public StatusEnum convertToEntityAttribute(String code) {
        return Stream.of(StatusEnum.values())
                .filter(c -> c.getCode().equals(code))
                .findFirst()
                .orElseThrow(IllegalArgumentException::new);
    }
}
